feat: fix comments
