+++ lib/app/features/user_block/model/database/blocked_users_database.c.dart
@@ -3,35 +3,19 @@
 import 'package:drift/drift.dart';
 import 'package:drift_flutter/drift_flutter.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:ion/app/exceptions/exceptions.dart';
 import 'package:ion/app/extensions/event_message.dart';
-import 'package:ion/app/features/auth/providers/auth_provider.c.dart';
 import 'package:ion/app/features/feed/notifications/data/database/converters/event_reference_converter.c.dart';
 import 'package:ion/app/features/ion_connect/database/converters/event_tags_converter.dart';
 import 'package:ion/app/features/ion_connect/ion_connect.dart';
 import 'package:ion/app/features/ion_connect/model/event_reference.c.dart';
 import 'package:ion/app/features/user_block/model/entities/blocked_user_entity.c.dart';
+import 'package:ion/app/features/user_block/providers/blocked_users_database_provider.c.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'blocked_users_database.c.g.dart';
 part 'dao/block_event_dao.c.dart';
 part 'tables/block_event_table.dart';
 part 'tables/deleted_block_event_table.dart';
-
-@Riverpod(keepAlive: true)
-BlockedUsersDatabase blockedUsersDatabase(Ref ref) {
-  final pubkey = ref.watch(currentPubkeySelectorProvider);
-
-  if (pubkey == null) {
-    throw UserMasterPubkeyNotFoundException();
-  }
-
-  final database = BlockedUsersDatabase(pubkey);
-
-  onLogout(ref, database.close);
-
-  return database;
-}
 
 @DriftDatabase(
   tables: [
